fix(transformedMeasure): added a layer over validateBoxModelMeasure in order to manage validations
